Support table valign (#56)
* feat: Support valign

* feat: Add valign writer

* feat: Add reader

* fix: indent

* 0.0.56

* fix: indent

* update

+++ docx-core/src/documents/elements/table_cell_width.rs
@@ -20,7 +20,7 @@ impl TableCellWidth {
 impl BuildXML for TableCellWidth {
     fn build(&self) -> Vec<u8> {
         XMLBuilder::new()
-            .table_cell_width(self.width, WidthType::DXA)
+            .table_cell_width(self.width as i32, WidthType::DXA)
             .build()
     }
 }
